+++ plotastrodata/plot_utils.py
@@ -4,7 +4,8 @@ import matplotlib.pyplot as plt
 from matplotlib.patches import Ellipse, Rectangle
 from dataclasses import dataclass
 
-from plotastrodata.other_utils import coord2xy, xy2coord, listing, estimate_rms
+from plotastrodata.coord_utils import coord2xy, xy2coord
+from plotastrodata.other_utils import listing, estimate_rms
 from plotastrodata.analysis_utils import AstroData, AstroFrame
